🏷️ Change int to double to match entity type for rateMin and rateMax params

+++ lib/features/tutee_assignments/domain/repositories/tutee_assignment_repo.dart
@@ -6,8 +6,8 @@ abstract class TuteeAssignmentRepo {
   Future<Either<Failure, List<TuteeAssignment>>> getByCriterion({
     Level level,
     Subject subject,
-    int rateMin,
-    int rateMax,
+    double rateMin,
+    double rateMax,
   });
 
   Future<Either<Failure, List<TuteeAssignment>>> getAssignmentList();
